breaking out the packageToEvaluate into its own file
